feat(avro): extract Apache Avro into its own module
