+++ modules/avro/src/avro-loader.ts
