+++ modules/avro/src/avro-schema-loader-types.ts
