+++ modules/avro/src/avro-stream.ts
