+++ modules/avro/src/avro-writer.ts
