+++ modules/avro/src/lib/encoders/encode-avro.ts
