+++ modules/avro/src/lib/parsers/parse-avro.ts
