+++ modules/avro/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": ["src/**/*"],
+  "exclude": ["node_modules"],
+  "compilerOptions": {"composite": true, "rootDir": "src", "outDir": "dist"},
+  "references": [{"path": "../compression"}, {"path": "../loader-utils"}, {"path": "../schema"}]
+}
